split collections into a separate package
