+++ src/Illuminate/Support/composer.json
@@ -18,7 +18,9 @@
         "ext-json": "*",
         "ext-mbstring": "*",
         "doctrine/inflector": "^1.1",
+        "illuminate/collections": "^8.0",
         "illuminate/contracts": "^8.0",
+        "illuminate/macroable": "^8.0",
         "nesbot/carbon": "^2.17",
         "voku/portable-ascii": "^1.4.8"
     },
